Add prev-next navigation if show_prev_next: true
Pages must be part of a list which have an 'order' attribute.
A link will be created to the page with an 'order' attribute of 1
less than the current page 'order', if such a page is in the list.
Also a link will be created to the page which has an order of 1
greater, if it exists.